requirements.txt
alembic==0.9.2
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
APScheduler==3.3.1
arrow==0.10.0
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
autopep8==1.3.1
autopy==0.51
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==2.0.0
blinker==1.4
BloomFilter==0.1.0
boto==2.48.0
bz2file==0.98
celery==3.1.25
celery-with-redis==3.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
CodeConvert==2.0.4
Cola==0.1.0
colorama==0.3.9
config==0.3.9
constantly==15.1.0
contextlib2==0.5.5
coverage==4.4.1
cryptography==2.0.2
csc-pysparse==1.1.1.4
csc-utils==0.6.7
cssselect==1.0.1
cycler==0.10.0
dateparser==0.6.0
decorator==4.1.2
defusedxml==0.5.0
demjson==2.2.4
Django==1.8
django-bootstrap-toolkit==2.15.0
django-filter==1.0.1
djangorestframework==3.5.4
djangorestframework-jwt==1.10.0
docutils==0.13.1
dominate==2.3.1
enum34==1.1.6
Flask==0.10.1
Flask-API==0.6.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Moment==0.5.1
flask-msearch==0.1.3
Flask-OpenID==1.2.5
Flask-PageDown==0.2.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.1.1
gensim==2.2.0
gevent==1.2.1
greenlet==0.4.12
grequests==0.3.0
gunicorn==19.1.1
h2==3.0.1
helper==2.4.2
hiredis==0.2.0
hpack==3.0.0
html==1.16
html5lib==0.999999999
hyperframe==5.1.0
hyperlink==17.3.0
idna==2.5
import==0.1.0
incremental==17.5.0
ipaddress==1.0.18
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
jdatetime==1.8.2
jieba==0.38
Jinja2==2.9.6
js2xml==0.2.2
kafka-python==1.3.3
kazoo==2.2.1
kombu==3.0.37
lxml==3.7.3
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
mechanize==0.3.3
mock==2.0.0
mongoengine==0.13.0
mysql==0.0.1
mysql-connector-python==2.1.6
MySQL-python==1.2.3
networkx==1.11
nltk==3.2.4
numpy==1.13.1+mkl
olefile==0.44
packaging==16.8
pandas==0.19.2
parsel==1.1.0
passlib==1.7.1
pathlib2==2.2.1
pbr==3.0.1
peewee==2.10.1
pickleshare==0.7.4
Pillow==4.1.1
ping==0.2
ply==3.10
priority==1.3.0
prompt-toolkit==1.0.14
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycodestyle==2.3.1
pycparser==2.18
pycurl==7.43.0
PyDispatcher==2.0.5
Pygments==2.2.0
PyJWT==1.5.2
pykafka==2.6.0
pymongo==3.4.0
PyMySQL==0.7.11
pyodbc==4.0.16
pyOpenSSL==17.2.0
pyparsing==2.2.0
pypiwin32==219
pyquery==1.2.17
pyrecsys==0.0.3
pyserial==3.4
pyspider==0.3.9
pytagcloud==0.3.5
pytesseract==0.1.6
python-cjson==1.2.1
python-dateutil==2.6.0
python-editor==1.0.3
python-etl==1.1
python-memcached==1.58
python-openid==2.2.5
python-recsys==0.2
pytz==2017.2
pywin32==221
queuelib==1.4.2
raven==6.1.0
readability-lxml==0.6.2
recsys==0.0.4
redis==2.10.5
redis-py-cluster==1.3.4
redisearch==0.6.1
regex==2017.4.29
requests==2.18.1
requests-oauth==0.4.1
retrying==1.3.3
rmtest==0.4.0
rq==0.8.0
rsa==3.4.2
ruamel.ordereddict==0.4.9
ruamel.yaml==0.14.12
scandir==1.5
scikit-learn==0.18.2
scikit-surprise==1.0.3
scipy==0.19.1
Scrapy==1.3.3
scrapy-crawlera==1.2.2
scrapy-proxies==0.3
scrapy-random-useragent==0.2
scrapy-redis==0.6.8
scrapy-splash==0.7.2
scrapyd==1.2.0
scrapyd-client==1.1.0
scrapyjs==0.2
selenium==3.4.2
service-identity==17.0.0
sha1==1.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
slimit==0.8.1
smart-open==1.5.3
SOAPpy==0.12.22
spynner==2.19
SQLAlchemy==1.1.4
tabulate==0.7.7
tblib==1.3.2
tornado==4.5.1
tqdm==4.11.2
traitlets==4.3.2
Twisted==17.5.0
typing==3.6.1
tzlocal==1.4
u-msgpack-python==2.4.1
umalqurra==0.2
urllib3==1.21.1
utils==0.9.0
vine==1.1.3
virtualenv==15.1.0
visitor==0.1.3
vulpo==1.2.1
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
Whoosh==2.7.4
win-unicode-console==0.5
wrapcache==1.0.8
WsgiDAV==2.2.2
wstools==0.4.5
WTForms==2.1
xlwt==1.2.0
zope.interface==4.4.2
